% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train:}{input data frame.}

\item{cl:}{true class value of your training data.}

\item{k_nn:}{integer representing the number of neighbors.}

\item{k_cv:}{integer representing the number of folds.}
}
\value{
class: a vector of the predicted class Ŷ i for all observation.

cv_err: a numeric with the cross-validation misclassification error.
}
\description{
Predict output class species using covariates
bill_length_mm, bill_depth_mm, flipper_length_mm, and body_mass_g.
}
\examples{
my_knn_cv(data_peng, test, k_nn = 1, k_cv = 5)

}
\keyword{prediction}
