% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fitting Linear Models}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula:}{a formula class object, similar to lm().}

\item{data:}{input data frame.}
}
\value{
table similar to the coefficient table from summary()
with rows for each coefficient (including the (Intercept)!)
and columns for the Estimate, Std. Error, t value, and Pr(>|t|).
There should be row and column names.
}
\description{
Used to fit linear models. It can be used to carry out regression,
single stratum analysis of variance and analysis of covariance.
}
\examples{
my_lm(mpg ~ hp + wt, data = mtcars)

}
\keyword{&}
\keyword{inference}
\keyword{prediction}
