% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric input indicates number of folds.}
}
\value{
a numeric representing the mean standard error across all the folds.
}
\description{
This function would use random forest algorithm to predict quantitive output
}
\examples{
my_rf_cv(5)
my_rf_cv(10)

}
\keyword{prediction}
