% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t.test}
\alias{my_t.test}
\title{t-Test Function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{alternative}{a character string specifying the alternative hypothesis.
This should only accept "two.sided", "less", or "greater".
Otherwise, your function should throw an informative error.}

\item{mu}{a number indicating the null hypothesis value of the mean.}
}
\value{
test_stat: the numeric test statistic.

df: the degrees of freedom.

alternative: the value of the \code{alternative}.

p_val: the numeric p-value.
}
\description{
Performs one and two sample t-tests on vectors of data.
}
\examples{
my_t.test(y, alternative = "two.sided", 0)
my_t.test(y, alternative = "lass", 0)

}
\keyword{inference}
