% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{Sample t-test Function}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{a numeric data vector}

\item{alternative}{the alternative hypothesis, in a character string}

\item{mu}{the null hypothesis mu}
}
\value{
A list with the following elements:
  test_stat: the numeric test statistic
  df: the degrees of freedom.
  alternative: the value of the parameter `alternative`
  p_val: the numeric p-value.
}
\description{
This function performs a one sample t-test.
}
\examples{
my_t_test(c(3.2, 5.5, 6.7, 6.7, 7.9, 8.4, 8.4, 8.4, 10.3, 11.2), "two.sided", 0)

}
\keyword{inference}
