% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function2_GIS_Louisiana.R
\name{plot_points}
\alias{plot_points}
\title{Generates points on a map of Louisiana}
\usage{
plot_points(coords, site_name)
}
\description{
Takes a .csv file with global coordinates in decimal degrees (longitude and latitude)
and plots them to a map of Louisiana's rivers
}
\details{
@param coords the read-in .csv file
 @param site_name the .csv file's site name column
 @return a map of Louisiana 
 @export plot_points
}
