% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function1.R
\name{query_genbank2}
\alias{query_genbank2}
\title{Queries data from genbank}
\usage{
query_genbank2(target, gene, return_max)
}
\arguments{
\item{target}{the taxa sequences are wanted for (can be species, genera, family level, etc)}

\item{gene}{the gene of interest -- such as CytB, ND2, COI, RAG2}

\item{return_max}{the number of sequences to return}
}
\value{
a list of sequences
}
\description{
Pulls data from genbank and complies in into a .fasta file for further anaylsis
}
