% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionOne.R
\name{query_id4}
\alias{query_id4}
\title{Queries data from genbank.}
\usage{
query_id4(taxalist)
}
\arguments{
\item{taxalist}{A list containing mulitple species and genus names you would like to query genbank for data for.}
}
\value{
megataxa A dataframe of all the species or genus you searched for and their associated data.
}
\description{
Each object listed in the list is searched inside genbank and all availble data
is converted into a dataframe containg an assortment of topics for each data input.
}
