% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visual_Data.R
\name{surveys_plot}
\alias{surveys_plot}
\title{Graphs filtered data from surveys}
\usage{
surveys_plot(output)
}
\arguments{
\item{output}{name of function used to filer data}
}
\value{
plot A visual of the filtered data
}
\description{
Graphs the average weight of surveys data by year
}
