% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Data.R
\name{surveys_weight_by_year}
\alias{surveys_weight_by_year}
\title{Filters data from surveys}
\usage{
surveys_weight_by_year(surveys)
}
\arguments{
\item{surveys}{A dataframe containing a column you would like to summarize}
}
\value{
weights A summary of the average weights
}
\description{
Filters the average weight of surveys data by year
}
