% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionTwo.R
\name{visualize_taxa}
\alias{visualize_taxa}
\title{Mapping queried data on a global map}
\usage{
visualize_taxa(penguin_df)
}
\arguments{
\item{penguin_df}{a data frame contianing scientificName, decimal latitude, and decimal longitude}
}
\description{
Data available in the table is filtered and mapped based of longitude and latitude on a global map that interactively changes based on your zoom.
}
