% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{Knn (K nearest neighbors) function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{input dataframe}

\item{cl}{values of true class from my training data}

\item{k_nn}{integer value represent the number of neighbors}

\item{k_cv}{integer value represent the number of folds}
}
\value{
a list with objects:
\code{class}, a vector of the predicted class yi_hat for all observations;
\code{cv_err}, a numeric with the cross-validation misclassification error.
}
\description{
Use Knn function to predict and record the errors
}
\examples{
my_penguins <- my_penguins \%>\% na.omit
my_knn_cv(my_penguins[3:6], my_penguins$species, 1, 5)

}
\keyword{prediction}
