% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame.}

\item{cl}{True class value of your training data.}

\item{k_nn}{Integer representing the number of neighbors.}

\item{k_cv}{Integer representing the number of folds.}
}
\value{
A list with objects:
\item{class}{A vector of the predicted class for all observations.}
\item{cv_err}{A numeric with the cross-validation misclassification error.}
}
\description{
This function implements \code{knn} to perform cross-validation.
}
\examples{
data <- na.omit(my_penguins)
data$species <- as.numeric(factor(data$species,
  levels = c('Adelie', 'Chinstrap', 'Gentoo'), labels = c(1, 2, 3)))
my_knn_cv(data[, 3:6], data[, 1], 1, 5)
my_knn_cv(data[, 3:6], data[, 1], 5, 5)

}
