% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{a data frame}

\item{cl}{class value of training data}

\item{k_nn}{integer of the number of neighbors}

\item{k_cv}{integer of the number of folds}
}
\value{
A list object of result output
}
\description{
This function predicts the class value cl by k-fold Cross-validation
         and k-nearest neighbors algorithms
}
\examples{
my_knn_cv(iris[, 1:4], iris$Species, 1, 5)
my_knn_cv(iris[, 1:4], iris$Species, 5, 5)

}
\keyword{prediction}
