% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{A data frame to be calculated for knn Cross-Validation.}

\item{cl}{A true class value of your data.}

\item{k_nn}{An integer representing the number of neighbors.}

\item{k_cv}{A integer representing the number of folds.}
}
\value{
A list with objects with vector of predicted class for observations
and numeric with the cross-validation misclassification error.
}
\description{
This function cauculated k-Nearest Neighbors Cross-Validation
}
\examples{

train <- my_gapminder[4]
cl <- unlist(my_gapminder[1])
my_knn_cv(train, cl, 5, 5)

}
\keyword{prediction}
