% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model Function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula class object, similar to lm().}

\item{data}{An input data frame.}
}
\value{
A fitted linear model.
}
\description{
This function fits a linear model to R.
}
\examples{
data(my_gapminder)
gapminderLm <- lm(pop ~ gdpPercap, data = my_gapminder)
summary(gapminderLm)
my_lm(gapminderLm, data = my_gapminder)

}
\keyword{Linear}
\keyword{Model}
