% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Leaner model Function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a formula of the class object}

\item{data}{input data frame}
}
\value{
a table with rows of each coefficient and columns of the estimate, standard error, test statistic and p-value
}
\description{
This function testing if there is a  coefficient between the model and data
}
\examples{
my_lm(bill_length_mm~bill_depth_mm, my_penguins)

}
\keyword{linear}
\keyword{model}
