% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula class object, similar to \code{lm()}.}

\item{data}{Input data frame.}
}
\value{
A table similar to the coefficent table from summary() with
rows for each coefficient (including the Intercept) and columns for the
estimate, standard error, t value, and \code{Pr(>|t|)}
}
\description{
This function fits a linear model.
}
\examples{
my_lm(lifeExp ~ gdpPercap + continent, my_gapminder)

}
