% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fitting a linear model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a formula class object}

\item{data}{input dataset}
}
\value{
a table with estimate, std error, tvalue and Pr>|t| of each coefficient
}
\description{
This function performs a linear regression model
}
\examples{
my_lm(mpg ~ hp + wt, data = mtcars)

}
\keyword{inference}
