% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{generate a function that fits a linear model in R}
\usage{
my_lm(f, data)
}
\arguments{
\item{f}{a \code{formula} class object, similar to lm().}

\item{data}{input data frame.}
}
\value{
A \code{table} similar to summary() :
with rows for each coefficient and columns for the
\code{Estimate}, \code{Std. Error}, \code{t value},
and \code{Pr(>|t|)}.
}
\description{
To create a functiona and this function fits a linear model in R.
}
\examples{

my_lm(lifeExp ~ gdpPercap, my_gapminder)


}
