% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross Validation Function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{An integer that represents the number of folds.}
}
\value{
A list of objects numerically with cross validation errors.
}
\description{
This function will predict output using given covariates.
}
\examples{
my_rf_cv(5)

}
\keyword{Forest}
\keyword{Random}
