% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Number of folds.}
}
\value{
A numeric with the cross-validation error
}
\description{
This function implements \code{randomForest} to perform cross-validation.
}
\examples{
my_rf_cv(2)
my_rf_cv(5)
my_rf_cv(10)

}
\keyword{cross-validation}
