% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross Validation Function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{integer of the number of folds}
}
\value{
a double of mean of MSE for all folds output error
}
\description{
This function predicts the class value cl by k-fold Cross-validation
         and random forest algorithms
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
