% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{generate a random forest cross-validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{a number for folds}
}
\value{
A \a numeric with the cross-validation error
}
\description{
we will predict output body_mass_g using covariates bill_length_mm,
bill_depth_mm, and flipper_length_mm.
}
\examples{

my_rf_cv(1)


}
\keyword{inference}
\keyword{prediction}
