% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-Test}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{The numeric vector of data.}

\item{alternative}{Character string specifying the alternative hypothesis.
This should only accept "two.sided", "less", or "greater".}

\item{mu}{A number indicating the null hypothesis value of the mean.}
}
\value{
A list with elements:
\item{test_stat}{The numeric test statistic.}
\item{df}{The degrees of freedom.}
\item{alternative}{The value of the parameter alternative.}
\item{p_val}{The numeric p-value.}
}
\description{
This function performs a one sample t-test.
}
\examples{
my_t.test(my_gapminder$lifeExp, "two.sided", 60)
my_t.test(my_gapminder$lifeExp, "less", 60)
my_t.test(my_gapminder$lifeExp, "greater", 60)

}
