% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{a one sample t-test in R}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{a numeric vector of data.}

\item{alternative}{a character string specifying the alternative hypothesis.
This should only accept \code{two.sided}, \code{less}, or \code{greater}.
Otherwise, your function should throw an informative error.}

\item{mu}{An integer representing the number of neighbors.}
}
\value{
A \code{list} with elements :
\code{test_stat}:the numeric test statistic
\code{df}:the degrees of freedom
\code{alternative}:the value of the parameter \code{alternative}
\code{p_val}:the numeric p-value
}
\description{
This function cauculated for a one sample t-test.
}
\examples{


my_t.test(my_gapminder$lifeExp, alternative = "two.sided", mu = 60)

}
\keyword{inference}
