% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t_test}
\alias{my_t_test}
\title{T test function}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{numeric vector of data}

\item{alternative}{hypothesis is that the true difference is different from zero, allow value is one of "two.sided", "greater" or "less"}

\item{mu}{the null hypothesis value of mean}
}
\value{
A list with the numeric test statistic, the degrees of freedom, the value of the parameter "alternative", and the p value
}
\description{
This function is used to compare the means of two groups
}
\examples{
my_t_test(rnorm(20), "two.sided", 1)
my_t_test(c(1:33), "greater", 5)

}
\keyword{hypothesis}
